requirements.txt
# [20251214_SECURITY] Security bump for CVE-2025-66416 (MCP DNS rebinding) and urllib3 redirect/decompression issues
astor
defusedxml  # [20251218_SECURITY] Safe XML parsing to prevent XXE attacks (B314)
graphviz
networkx
z3-solver
requests
tqdm
pyyaml
crewai
langchain
langgraph  # [20251217_FEATURE] LangGraph integration for autonomous fixing
autogen
anthropic
openai
antlr4-python3-runtime
toml
plotly
matplotlib
esprima
javalang
tree-sitter
tree-sitter-java
tree-sitter-javascript  # [20251218_BUGFIX] Add missing JavaScript tree-sitter binding
tree-sitter-typescript  # [20251218_BUGFIX] Add missing TypeScript tree-sitter binding
mcp>=1.23.0
pydantic>=2.11.0  # [20251230_BUGFIX] Required by mcp>=1.25.0
pydantic-settings>=2.5.2  # [20251230_BUGFIX] Required by mcp>=1.25.0
urllib3>=2.6.0
uvicorn
flask>=2.0.0  # [20251218_BUGFIX] Add missing Flask dependency from pyproject.toml
pytest-asyncio>=0.23  # [20251214_TEST] Enable async test execution
pytest-timeout>=2.3.1  # [20251230_TEST] Fail fast on hangs
freezegun>=1.4.0  # [20251228_TEST] Time-mocking for expiration/grace tests
structlog>=24.1.0  # [20251216_FEATURE] v2.2.0 Structured MCP logging
PyJWT[crypto]>=2.10.1  # [20251230_BUGFIX] Align with mcp>=1.25.0
cryptography>=41.0.0  # [20251225_FEATURE] v3.3.0 RSA key support for JWT
